test: rename coverage-named test files to feature-based names
Rename test files from 'coverage_<area>' to their domain name to
reflect that they test real behavior rather than serving as
coverage hacks:

- test_coverage_binary_sensor.py -> test_binary_sensor.py
- test_coverage_diagnostics.py -> test_diagnostics.py
- test_coverage_repair_flow.py -> test_repair_flow.py
- test_coverage_select.py -> test_select.py
- test_coverage_retry_utils.py -> test_retry_utils.py

Merge test_coverage_mapping_transform.py into the existing
test_mapping_transform.py (which already covered signed conversion)
so mapping_transform behavior lives in a single test file.
