+++ tests/integration/test_diagnostics.py
